move some file
